pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flammap"
version = "0.1.0"
description = "Config-driven pipeline for generating flammability limits and phase diagrams."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ase",
  "cantera",
  "matplotlib",
  "molsym",
  "numpy",
  "pandas",
  "PyYAML",
  "scipy",
]

[project.scripts]
flammap = "flammability.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
